Create new file with the table column helper and components
